+++ mlx_vlm/models/laguna/__init__.py
@@ -1,0 +1,4 @@
+from .config import ModelConfig
+from .laguna import Model
+
+__all__ = ["Model", "ModelConfig"]
