+++ mlx_vlm/models/laguna/language.py
@@ -1,5 +1,4 @@
-from dataclasses import dataclass, field
-from typing import Any, Dict, List, Optional, Union
+from typing import Any, Dict, Optional, Union
 
 import mlx.core as mx
 import mlx.nn as nn
@@ -7,150 +6,13 @@ from mlx_lm.models.activations import swiglu
 from mlx_lm.models.rope_utils import initialize_rope
 from mlx_lm.models.switch_layers import SwitchGLU
 
-from .base import (
-    BaseModelConfig,
-    InputEmbeddingsFeatures,
+from ..base import (
     LanguageModelOutput,
     create_attention_mask,
     scaled_dot_product_attention,
 )
-from .cache import KVCache, RotatingKVCache
-
-
-@dataclass
-class ModelConfig(BaseModelConfig):
-    model_type: str
-    vocab_size: int
-    hidden_size: int
-    intermediate_size: int
-    num_hidden_layers: int
-    num_attention_heads: int
-    num_key_value_heads: int
-    head_dim: int
-    max_position_embeddings: int
-    rms_norm_eps: float = 1e-6
-    qkv_bias: bool = False
-    attention_bias: bool = False
-    gating: Union[bool, str] = True
-    tie_word_embeddings: bool = False
-    rope_theta: float = 500000.0
-    rope_parameters: Optional[Dict[str, Any]] = None
-    rope_scaling: Optional[Dict[str, Any]] = None
-    partial_rotary_factor: Optional[float] = None
-    rope_style: str = "rotate-half"
-    sliding_window: Optional[int] = None
-    layer_types: Optional[List[str]] = None
-    num_attention_heads_per_layer: Optional[List[int]] = None
-    swa_rope_parameters: Optional[Dict[str, Any]] = None
-    swa_attention_sink_enabled: bool = False
-    num_experts: int = 0
-    num_experts_per_tok: int = 0
-    moe_intermediate_size: int = 0
-    shared_expert_intermediate_size: int = 0
-    norm_topk_prob: bool = True
-    decoder_sparse_step: int = 1
-    mlp_only_layers: List[int] = field(default_factory=lambda: [0])
-    mlp_layer_types: Optional[List[str]] = None
-    moe_routed_scaling_factor: float = 1.0
-    moe_apply_router_weight_on_input: bool = False
-    moe_router_logit_softcapping: float = 0.0
-    moe_router_use_sigmoid: bool = True
-    bos_token_id: Optional[int] = None
-    eos_token_id: Optional[Union[int, List[int]]] = None
-    pad_token_id: Optional[int] = None
-
-    def __post_init__(self):
-        if self.gating is True:
-            self.gating = "per-head"
-
-        if self.layer_types is None:
-            self.layer_types = ["full_attention"] * self.num_hidden_layers
-        if len(self.layer_types) != self.num_hidden_layers:
-            raise ValueError("layer_types must match num_hidden_layers.")
-
-        if self.mlp_layer_types is not None:
-            if len(self.mlp_layer_types) != self.num_hidden_layers:
-                raise ValueError("mlp_layer_types must match num_hidden_layers.")
-            self.mlp_only_layers = [
-                idx
-                for idx, layer_type in enumerate(self.mlp_layer_types)
-                if layer_type == "dense"
-            ]
-
-        if self.num_attention_heads_per_layer is None:
-            self.num_attention_heads_per_layer = [
-                self.num_attention_heads
-            ] * self.num_hidden_layers
-        if len(self.num_attention_heads_per_layer) != self.num_hidden_layers:
-            raise ValueError(
-                "num_attention_heads_per_layer must match num_hidden_layers."
-            )
-        if any(
-            h % self.num_key_value_heads for h in self.num_attention_heads_per_layer
-        ):
-            raise ValueError(
-                "Every query-head count must be divisible by num_key_value_heads."
-            )
-
-        rope_parameters = (
-            dict(self.rope_parameters)
-            if self.rope_parameters is not None
-            else (
-                dict(self.rope_scaling)
-                if self.rope_scaling is not None
-                else {"rope_type": "default", "rope_theta": self.rope_theta}
-            )
-        )
-
-        layer_types = set(self.layer_types)
-        layer_rope_parameters = {
-            k: v
-            for k, v in rope_parameters.items()
-            if k in layer_types and isinstance(v, dict)
-        }
-        if layer_rope_parameters:
-            top_level_parameters = {
-                k: v
-                for k, v in rope_parameters.items()
-                if k not in layer_types and not isinstance(v, dict)
-            }
-
-            def rope_parameters_for(layer_type: str) -> Dict[str, Any]:
-                params = dict(layer_rope_parameters.get(layer_type, {}))
-                for k, v in top_level_parameters.items():
-                    params.setdefault(k, v)
-                return params
-
-            default_layer_type = (
-                "full_attention"
-                if "full_attention" in layer_rope_parameters
-                else next(iter(layer_rope_parameters))
-            )
-            self.rope_parameters = rope_parameters_for(default_layer_type)
-
-            if (
-                self.swa_rope_parameters is None
-                and "sliding_attention" in layer_rope_parameters
-            ):
-                self.swa_rope_parameters = rope_parameters_for("sliding_attention")
-        else:
-            self.rope_parameters = rope_parameters
-
-        if self.swa_rope_parameters is not None:
-            self.swa_rope_parameters = dict(self.swa_rope_parameters)
-
-        self.rope_parameters.setdefault("rope_type", "default")
-        if self.swa_rope_parameters is not None:
-            self.swa_rope_parameters.setdefault("rope_type", "default")
-
-        if self.partial_rotary_factor is not None:
-            self.rope_parameters.setdefault(
-                "partial_rotary_factor", self.partial_rotary_factor
-            )
-            if self.swa_rope_parameters is not None:
-                self.swa_rope_parameters.setdefault(
-                    "partial_rotary_factor", self.partial_rotary_factor
-                )
+from ..cache import KVCache, RotatingKVCache
+from .config import ModelConfig
 
 
 def _rope_base(args: ModelConfig, rope_config: Dict[str, Union[float, str]]) -> float:
@@ -406,7 +268,7 @@ class LagunaModel(nn.Module):
         return self.norm(h)
 
 
-class _LanguageModel(nn.Module):
+class LanguageModel(nn.Module):
     def __init__(self, args: ModelConfig):
         super().__init__()
         self.args = args
@@ -545,64 +407,3 @@ class _LanguageModel(nn.Module):
             )
             for layer in self.layers
         ]
-
-
-class Model(nn.Module):
-    def __init__(self, config: ModelConfig):
-        super().__init__()
-        self.config = config
-        self.model_type = config.model_type
-        self.language_model = _LanguageModel(config)
-
-    def get_input_embeddings(
-        self,
-        input_ids: Optional[mx.array] = None,
-        pixel_values: Optional[mx.array] = None,
-        **kwargs,
-    ) -> InputEmbeddingsFeatures:
-        return InputEmbeddingsFeatures(
-            inputs_embeds=self.language_model.model.embed_tokens(input_ids)
-        )
-
-    def __call__(
-        self,
-        input_ids: mx.array,
-        pixel_values: mx.array = None,
-        mask: mx.array = None,
-        cache=None,
-        **kwargs,
-    ) -> LanguageModelOutput:
-        input_embeddings_features = self.get_input_embeddings(input_ids, pixel_values)
-        return self.language_model(
-            input_ids,
-            cache=cache,
-            inputs_embeds=input_embeddings_features.inputs_embeds,
-            **kwargs,
-        )
-
-    def sanitize(self, weights):
-        weights = self.language_model.sanitize(weights)
-
-        def transform_key(key):
-            if key.startswith("language_model."):
-                return key
-            if key.startswith("model.") or key.startswith("lm_head."):
-                return f"language_model.{key}"
-            return key
-
-        return {transform_key(k): v for k, v in weights.items()}
-
-    @property
-    def quant_predicate(self):
-        return self.language_model.quant_predicate
-
-    @property
-    def cast_predicate(self):
-        return self.language_model.cast_predicate
-
-    @property
-    def layers(self):
-        return self.language_model.layers
-
-    def make_cache(self):
-        return self.language_model.make_cache()
